% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{estcondmin}
\alias{estcondmin}
\title{Function to optimize}
\usage{
estcondmin(y, X, lambda = 0)
}
\arguments{
\item{y}{A vector that gives the predict variable}

\item{X}{A matrix that gives the samples of features}

\item{lambda}{A number to penalize the complexity of the model}
}
\value{
A coeficient of selected feature form the sufficient condition.
}
\description{
This is a function of estimate the coeficients of some sufficient condition
from the dataset.
}
\examples{
d = gen_dat()
estcondmin(d$y, d$X, lambda = 0.1)

}
