% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{estcondmin2}
\alias{estcondmin2}
\title{Function 2 to optimize}
\usage{
estcondmin2(y, X, lambda = 0, epsilon = 0.01)
}
\arguments{
\item{y}{A vector that gives the predict variable}

\item{X}{A matrix that gives the samples of features}

\item{lambda}{A number to penalize the complexity of the model}

\item{epsilon}{A muber of disturb to the central point}
}
\value{
A coeficient of selected feature form the sufficient condition.
}
\description{
This is a function of estimate the coeficients of some sufficient condition
from the dataset.
}
\examples{
d = gen_dat()
estcondmin(d$y, d$X, lambda = 0.1)

}
