% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{gen_dat}
\alias{gen_dat}
\title{Data Generate}
\usage{
gen_dat(n = 100, beta = c(1, 1))
}
\arguments{
\item{n}{A integer that gives the sample size}

\item{beta}{A vector gives the coeficients of the true model}
}
\value{
This function returns a list of two element \code{y} and \code{X}
}
\description{
This a a simple function to create some test data for our model.
 The model generate the dat set is that: \deqn{y = \sum_{i = 0}^p \beta_i x_i  + e }{ y = \sum_{i = 0}^p \beta_i x_i  + e }
}
\examples{
gen_dat(n = 100, beta = c(1,1))

}
